Move python code to a library
