+++ python_backend/README.md
@@ -1,0 +1,3 @@
+This has a dependency on [haversine](https://pypi.org/project/haversine/)
+
+TODO: Make this a proper Python library and consume in the frontend as a wheel?
