+++ template_app/static/backend.py
@@ -1,0 +1,1 @@
+../../python_backend/backend.py
